fix(config): FE-00 Allow functions returning promises to be passed into jsx attributes that expect void return
A recent change in the `no-misused-promises` rule made it fail when we
pass a function returning a promise into a prop that expects a void
return. This removes the check and relies on TS to make sure promises
are handled correctly.

```ts
const clicked = async () => {};

return <Component onClick={clicked} />
```

It won't allow promises in methods that are explicitly typed as void:

```diff
-const method = (): void => fetch();
+const method = (): void => {
+  void fetch();
+}
```